+++ frontend/src/components/Flats/FlatsDashboard.js
@@ -1,10 +1,10 @@
 import React from "react";
 
 const FlatsDashboard = () => {
-    return (
-        <>
-        <h3>Hello welcome...</h3>
-        </>
-    )
-}
+  return (
+    <>
+      <h3>Hello welcome...</h3>
+    </>
+  );
+};
 export default FlatsDashboard;
