--- frontend/src/components/Home.js
@@ -1,22 +1,0 @@
-import React, { useEffect, useState } from "react";
-import axios from "axios";
-
-const Home = () => {
-  const [message, setMessage] = useState("");
-
-  useEffect(() => {
-    axios
-      .get("http://127.0.0.1:8000/api/hello/")
-      .then((response) => setMessage(response.data.message))
-      .catch((error) => console.error(error));
-  }, []);
-
-  return (
-    <div className="container mt-5">
-      <h1>Welcome to the Home Page</h1>
-      <p>{message}</p>
-    </div>
-  );
-};
-
-export default Home;
