--- frontend/src/components/SignUp.js
@@ -1,40 +1,0 @@
-import React, { useState } from "react";
-import axios from "axios";
-
-const Signup = () => {
-  const [mobile, setMobile] = useState("");
-  const [password, setPassword] = useState("");
-
-  const handleSignup = async (e) => {
-    e.preventDefault();
-    try {
-      await axios.post("http://127.0.0.1:8000/users/signup/", {
-        mobile,
-        password,
-      });
-      alert("User created successfully!");
-    } catch (error) {
-      alert("Error: Mobile number may already exist");
-    }
-  };
-
-  return (
-    <form onSubmit={handleSignup}>
-      <input
-        type="text"
-        placeholder="Mobile"
-        value={mobile}
-        onChange={(e) => setMobile(e.target.value)}
-      />
-      <input
-        type="password"
-        placeholder="Password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <button type="submit">Sign Up</button>
-    </form>
-  );
-};
-
-export default Signup;
